Credit
(C) Takes user input for a credit card number and based on a subset of card types tells the user if it is a valid Visa, Mastercard, American Express card, or if it is invalid.  Subset of card number types defined for output results are: American Express 15 digits starting with 34 or 37, Mastercard 16 digits starting with 51-55, Visa 13 or 16 digits starting with 4, anything else is invalid.